// gestionDeCompteBancaireApp/src/app/model/client.ts
export class Client {
    idClient?: number;
    numClient?: string;
    nom?: string;
    mail?: string;
    numTel?: string;
    prenom?: string;
    adressePostale?: string;
    dateNaissance?: Date;
    lieuNaissance?: string;
    nationalite?: string;
    profession?: string;
    revenu?: string;
    solde?:number;
    idBanquier?: number;
}